Update for 1.12.2
Note that Spigot is named 1.12.20 because in the last version I called it 1.12.10 and the way it's set up it has to be >.10 for Synology to update properly.  Should be able to just live with it until next major MC update.  Other option seems to require people to uninstall Spigot.