Update to 1.15.2
I'm a terrible human being for neglecting something as important as MC for so long, but I've updated my repository and the package server for 1.15.2.  I'm going to take a look at the redstone server that's out there next, as someone's sent a pull request that may already have it working.  Finally: it still deletes your JSON files, so backup first, folks!